refactor: split dashboard sections + extract OTP flow + draft storage helper
